new control algorithm

+++ Core/Inc/Servo.h
@@ -15,7 +15,7 @@
 #include "stm32f1xx_hal.h"
 
 
-void setServoAngle(TIM_HandleTypeDef* timer, uint32_t channel, uint16_t period, int16_t angle);
+void setServoAngle(TIM_HandleTypeDef* timer, uint32_t channel, uint16_t period, int8_t angle);
 void setServoPWM(TIM_HandleTypeDef* timer, uint32_t channel, uint16_t period, uint16_t pulse);
 
 #endif /* INC_SERVO_H_ */
